Pakeistas VerbalMemoryManagerTests, kad sutaptu su kitais pavadinimas, prideti NumberMemoryManager testai.
